D3,Sravanthi| #475 | Move identifier to api, renamed variables
